- Bunch of significant changes.   * Didn't break up into smaller commit since this fixes caused other problems and wanted no new broken unit tests - Local Thresholding now refers to width instead of radius - Can specify a local thresholding region using ConfigLength - Created Otsu2

+++ main/boofcv-ip/src/main/java/boofcv/struct/ConfigLength.java
@@ -42,7 +42,7 @@ public class ConfigLength implements Configuration {
 	public ConfigLength() {
 	}
 
-	public static ConfigLength fixed(int pixels ) {
+	public static ConfigLength fixed(double pixels ) {
 		return new ConfigLength(pixels,-1);
 	}
